Update get_endianness function

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -9,8 +9,8 @@
 
 int get_endianness(void)
 {
-	unsigned int x = 1;
+	int x = 1;
 	char *c = (char *)&x;
 
-	return (int)(*c);
+	return (*c);
 }
